requirements.txt
sentence-transformers
pinecone
python-dotenv
groq
streamlit
pypdf